+++ src/main/java/com/mageddo/dnsproxyserver/config/entrypoint/ConfigJson.java
@@ -40,4 +40,6 @@ public interface ConfigJson {
   URI getDockerHost();
 
   Boolean getResolvConfOverrideNameServers();
+
+  Boolean getNoRemoteServers();
 }