+++ src/main/java/com/mageddo/dnsproxyserver/quarkus/Instances.java
@@ -1,12 +1,15 @@
 package com.mageddo.dnsproxyserver.quarkus;
 
+import com.mageddo.dnsproxyserver.di.InstanceImpl;
+
 import javax.enterprise.inject.Instance;
 import java.util.List;
 import java.util.Set;
 import java.util.stream.Collectors;
+import java.util.stream.Stream;
 
 public class Instances {
-  public static  <T> List<T> toList(Instance<T> instance) {
+  public static <T> List<T> toList(Instance<T> instance) {
     return instance
       .stream()
       .toList()
@@ -18,5 +21,9 @@ public class Instances {
       .stream()
       .collect(Collectors.toSet())
       ;
+  }
+
+  public static <T> Instance<T> of(T... instances) {
+    return new InstanceImpl<>(Stream.of(instances).toList());
   }
 }